Add script to emulate Projectile functionality
This script searches for all git repositories in code/ and its children
and gives a list with fuzzy search for the user to choose from.
Once a project is selected, a new session is created with nvim open.

TODO: Ensure the new terminals being open are open in the correct
directory